refactor: modularize api and utils
