+++ server.ts
@@ -6,6 +6,9 @@ import {
   join,
 } from "https://deno.land/std@0.224.0/path/mod.ts";
 import cheerio from "npm:cheerio@1.0.0-rc.12";
+import { parseArticles, randomSentence } from "./articleUtils.js";
+import { scrapeWx } from "./api/wx.js";
+import { scrapeBil } from "./api/bil.js";
 
 const CORS_HEADERS = {
   "Access-Control-Allow-Origin": "*",
@@ -49,26 +52,6 @@ const swHtml = `const IMG_CACHE = ${JSON.stringify(cacheImgDomain)};\n${swRaw}`;
 const adminHtml = injectConfig(
   await Deno.readTextFile(join(__dirname, "admin.html")),
 );
-const fallbackSentences = [
-  "小荷才露尖尖角",
-  "早有蜻蜓立上头",
-  "采菊东篱下",
-  "悠然见南山",
-  "看看内容吧",
-  "日长篱落无人过", 
-  "惟有蜻蜓蛱蝶飞",
-  "小娃撑小艇",
-  "日长篱落无人过", 
-  "惟有蜻蜓蛱蝶飞",
-  "偷采白莲回",
-  "不解藏踪迹", 
-  "浮萍一道开",
-];
-function randomSentence() {
-  return fallbackSentences[
-    Math.floor(Math.random() * fallbackSentences.length)
-  ];
-}
 // 微信文章列表
 const WX_URL = Deno.env.get("WX_URL") || "article.txt";
 const DAILY_URL = "https://www.cikeee.com/api?app_key=pub_23020990025";
@@ -84,47 +67,6 @@ interface ArticleMeta {
   date?: string;
 }
 
-function parseArticles(text: string): ArticleMeta[] {
-  const trimmed = text.trim();
-  if (!trimmed.startsWith("---")) {
-    return trimmed
-      .split(/\r?\n/)
-      .map((l) => ({ url: l.trim() }))
-      .filter((a) => a.url);
-  }
-
-  const parts = trimmed.split(/^---\s*$/m).map((p) => p.trim()).filter(Boolean);
-  const articles: ArticleMeta[] = [];
-  for (const part of parts) {
-    const lines = part.split(/\r?\n/);
-    const meta: any = {};
-    let current: string | null = null;
-    for (const line of lines) {
-      const kv = line.match(/^([\w]+):\s*(.*)$/);
-      if (kv) {
-        current = kv[1];
-        const value = kv[2];
-        if (value === "") {
-          if (current === "tags") {
-            meta[current] = [];
-          } else {
-            meta[current] = "";
-          }
-        } else {
-          meta[current] = value;
-        }
-        continue;
-      }
-      const m = line.match(/^\s*-\s*(.+)$/);
-      if (m && current) {
-        if (!Array.isArray(meta[current])) meta[current] = [];
-        meta[current].push(m[1]);
-      }
-    }
-    if (meta.url) articles.push(meta as ArticleMeta);
-  }
-  return articles;
-}
 
 let articles: ArticleMeta[] = [];
 try {
@@ -191,99 +133,6 @@ let wxCache: { data: unknown; timestamp: number } = { data: null, timestamp: 0 }
 let bilCache: { data: unknown; timestamp: number } = { data: null, timestamp: 0 };
 
 // ---------- 业务函数 ----------
-async function scrape(article: ArticleMeta) {
-  const { url } = article;
-  const controller = new AbortController();
-  const timer = setTimeout(() => controller.abort(), 15_000);
-
-  try {
-    const res = await fetch(url, {
-      headers: { "User-Agent": "Mozilla/5.0 (Deno)" },
-      signal: controller.signal,
-    });
-    const html = await res.text();
-
-    const $ = cheerio.load(html, { decodeEntities: false });
-
-    const name = article.title ||
-      $("#activity-name").text().trim() ||
-      $(".rich_media_title").text().trim() ||
-      randomSentence();
-
-    const time = article.date ||
-      $("#publish_time").text().trim() ||
-      $('meta[property="article:published_time"]').attr("content")?.trim();
-
-    const description = article.describe ||
-      $('meta[property="og:description"]').attr("content")?.trim() ||
-      $("#js_content p").first().text().trim();
-
-    const images: string[] = [];
-    $("#js_content img").each((_, el) => {
-      const src = $(el).attr("data-src") || $(el).attr("src");
-      if (src) images.push(src.split("?")[0]);
-    });
-
-    // 解析 <catch id="json-wx">
-    const jsonWxRaw = $("catch#json-wx").html()?.trim();
-    let jsonWx: unknown;
-    if (jsonWxRaw) {
-      try {
-        jsonWx = JSON.parse(jsonWxRaw.replace(/&quot;/g, '"'));
-      } catch (e) {
-        jsonWx = { parseError: e.message, raw: jsonWxRaw };
-      }
-    }
-
-    return {
-      [name]: {
-        time,
-        description,
-        images,
-        jsonWx,
-        url,
-        tags: article.tags,
-        abbrlink: article.abbrlink,
-        date: article.date,
-      },
-    };
-  } finally {
-    clearTimeout(timer);
-  }
-}
-
-async function scrapeBili(article: ArticleMeta) {
-  const { url } = article;
-  const controller = new AbortController();
-  const timer = setTimeout(() => controller.abort(), 15_000);
-  try {
-    const res = await fetch(url, {
-      headers: { "User-Agent": "Mozilla/5.0 (Deno)" },
-      signal: controller.signal,
-    });
-    const html = await res.text();
-    const $ = cheerio.load(html, { decodeEntities: false });
-    const name = article.title || $(".opus-module-title__text").first().text().trim() || randomSentence();
-    const description = article.describe || $(".opus-module-content").first().text().trim();
-    const images: string[] = [];
-    $(".opus-module-content img").each((_, el) => {
-      const src = $(el).attr("src");
-      if (src) images.push(src.startsWith("//") ? `https:${src}` : src);
-    });
-    return {
-      [name]: {
-        description,
-        images,
-        url,
-        tags: article.tags,
-        abbrlink: article.abbrlink,
-        date: article.date,
-      },
-    };
-  } finally {
-    clearTimeout(timer);
-  }
-}
 
 // ---------- 工具：把文章 HTML 里的微信图片替换成代理地址 (可选) ----------
 function proxifyHtml(html: string): string {
@@ -457,7 +306,9 @@ async function handler(req: Request): Promise<Response> {
       }
 
       const wxArticles = articles.filter((a) => a.url.includes("mp.weixin.qq.com"));
-      const results = await Promise.allSettled(wxArticles.map(scrape));
+      const results = await Promise.allSettled(
+        wxArticles.map((a) => scrapeWx(a, cheerio, randomSentence))
+      );
       const merged: Record<string, unknown> = {};
       results.forEach((r, i) => {
         if (r.status === "fulfilled") {
@@ -481,7 +332,9 @@ async function handler(req: Request): Promise<Response> {
       }
 
       const bilArticles = articles.filter((a) => a.url.includes("bilibili.com"));
-      const results = await Promise.allSettled(bilArticles.map(scrapeBili));
+      const results = await Promise.allSettled(
+        bilArticles.map((a) => scrapeBil(a, cheerio, randomSentence))
+      );
       const merged: Record<string, unknown> = {};
       results.forEach((r, i) => {
         if (r.status === "fulfilled") {
